Correct the pagination logic
