+++ src/index.html
@@ -70,6 +70,6 @@
         </div>
       </section>
     </main>
-    <script src="./js/index.js"></script>
+    <script src="../dist/index.js"></script>
   </body>
 </html>
