+++ src/js/pagination.js
@@ -1,21 +1,40 @@
-const paginationUl = document.querySelector('.pagination');
+import { getScores } from './data';
+
 let current = 1;
-function pages(pages, paginationUl) {
+
+export const display = async (data) => {
+  const tableBody = document.querySelector('.table-body');
+  const paginationUl = document.querySelector('.pagination');
+  const paginatedItems = paginate(current, 4, data);
+
+  tableBody.innerHTML = '';
+  paginatedItems.forEach((player) => {
+    const tr = document.createElement('tr');
+    tr.innerHTML = `<td>${player.user} : ${player.score}</td>`;
+    tableBody.appendChild(tr);
+  });
+
+  const paginationSize = Math.ceil(data.length / 4);
+  pages(paginationSize, paginationUl);
+};
+
+export function pages(pages, paginationUl) {
   paginationUl.innerHTML = '';
   for (let i = 1; i <= pages; i += 1) {
     const paginationLi = document.createElement('li');
     paginationLi.className =
       'p-item rounded-circle text-center d-flex justify-content-center';
     paginationLi.id = i;
-    if (current === paginationLi.id) {
-      paginationLi.classList.add('bg-dark');
+    if (current == paginationLi.id) {
+      paginationLi.classList.add('bg-primary');
       paginationLi.classList.add('p-active');
-    } else {
-      paginationLi.classList.remove('active');
     }
     // eslint-disable-next-line no-loop-func
-    paginationLi.addEventListener('click', () => {
+    paginationLi.addEventListener('click', async () => {
       current = paginationLi.id;
+      const result = await getScores();
+      const { result: data } = await result.json();
+      display(data);
     });
     paginationLi.innerHTML = `<a class='p-link' href='#'>${i}
     </a>`;
@@ -29,3 +48,5 @@ function paginate(currentPage = 1, rows, array) {
   const paginatedItems = array.slice(loopStart, loopStart + rows);
   return paginatedItems;
 }
+
+export default paginate;
